Menu V2.0 Buttons working + Labels, missing music before completed

+++ include/states/EndState.h
@@ -9,7 +9,7 @@ public:
     EndState(Game* gameObj);
     void Enter() override;
     void Update() override;
-    void UpdatePhysics(int dt) override;
+    void UpdatePhysics(double dt) override;
     void Draw() override;
     void Exit() override;
     std::string getDescription() override;
